automation/requirements.txt
# Longevity Knowledge Platform - Content Pipeline
# Install: pip install -r requirements.txt

# ============ Core ============
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# ============ AI Providers ============
# Gemini (Google)
google-generativeai>=0.5.0

# OpenAI / Kimi (OpenAI-compatible)
openai>=1.12.0

# ============ Document Processing ============
# Docling for PDF processing (optional - heavy dependency)
# docling>=1.0.0

# ============ Data Processing ============
pandas>=2.0.0
numpy>=1.24.0

# ============ Web & API ============
requests>=2.31.0
httpx>=0.26.0

# ============ Utilities ============
pydantic>=2.5.0
tenacity>=8.2.0  # Retry logic
rich>=13.0.0     # Pretty console output
